Organize my files insode ./src
