+++ src/components/CardActions.js
@@ -2,3 +2,5 @@
 function CardActions(){
 
 }
+
+export default CardActions;
