+++ src/components/Cards.js
@@ -2,3 +2,5 @@
 function Cards(){
 
 }
+
+export default Cards();
