+++ src/components/ColumnContainer.js
@@ -2,3 +2,5 @@
 function ColumnContainer(){
 
 }
+
+export default ColumnContainer();
