+++ src/components/ColumnListContainer.js
@@ -2,3 +2,5 @@
 function ColumnListContainer(){
 
 }
+
+export default ColumnListContainer();
